requirements.txt
streamlit
openai
pandas
scikit-learn
xgboost
shap
fpdf
python-dotenv